Builder stage popping several payloads at a time
